wine-model/pyproject.toml
[project]
name = "wine-model"
version = "0.1.0"
description = ""
authors = [
    {name = "Luiz Rath",email = "44349218+martianplatypus@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit (>=1.46.0,<2.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pytest (>=8.4.1,<9.0.0)"
]

[tool.poetry]
packages = [{include = "wine_model", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
